requirements.txt
aiohttp-jinja2==1.6
cachetools==6.2.0
cryptography==45.0.4
Flask==3.1.2
humanize==4.13.0
parse-torrent-title==2.8.1
pymongo==4.14.1
python-dotenv==1.1.1
python-telegram-bot==22.3
requests-html==0.10.0
schedule==1.2.2
TgCrypto==1.2.5
themoviedb==1.0.2
colorama==0.4.6
requests==2.32.4
RapidFuzz==3.14.1
d4rktg==1.5.1
psutil==7.1.0
aiohttp-session
python-multipart
fastapi==0.121.1
uvicorn==0.38.0
pytz==2025.2
google-auth
google-auth-oauthlib
google-auth-httplib2